[package] Add raise_packaging_errors flag to PackageExporter
**Summary**
This commit adds a `raise_packaging_errors` flag to `PackageExporter`
(defaults to `True`). If set to `False`, then `PackageExporter` does not
raise an exception, and instead records all of the modules it could not
save into the package into `self.broken_modules` along with the reason.

A `PackageExporter` with `raise_packaging_errors=True` will be used in
follow-up PRs as a basis for a dependency exploration tool.

**Test Plan**
This commit adds a unit test to `TestMisc` that tests this new
behaviour.

ghstack-source-id: 9afadca5167e293114beb0a752b5afc86d9321c1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56234